Moving tests into their own assembly
Including nunit.framework in lib dir for mono
