+++ tools/Tests/Tests.csproj
@@ -1,0 +1,102 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="3.5" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProductVersion>9.0.21022</ProductVersion>
+    <SchemaVersion>2.0</SchemaVersion>
+    <ProjectGuid>{0176E5C5-C12B-4386-8F23-4F4AD7462519}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>Tests</RootNamespace>
+    <AssemblyName>Tests</AssemblyName>
+    <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
+    <FileAlignment>512</FileAlignment>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="nunit.framework, Version=2.5.8.10295, Culture=neutral, PublicKeyToken=96d09a1eb7f44a77, processorArchitecture=MSIL">
+      <HintPath>..\lib\nunit.framework.dll</HintPath>
+    </Reference>
+    <Reference Include="System" />
+    <Reference Include="System.Core">
+      <RequiredTargetFramework>3.5</RequiredTargetFramework>
+    </Reference>
+    <Reference Include="System.Xml.Linq">
+      <RequiredTargetFramework>3.5</RequiredTargetFramework>
+    </Reference>
+    <Reference Include="System.Data.DataSetExtensions">
+      <RequiredTargetFramework>3.5</RequiredTargetFramework>
+    </Reference>
+    <Reference Include="System.Data" />
+    <Reference Include="System.Xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="Application.cs" />
+    <Compile Include="BuildTest.cs" />
+    <Compile Include="CssProcessor.cs" />
+    <Compile Include="DependencyMapTest.cs" />
+    <Compile Include="HtmlElementNames.cs" />
+    <Compile Include="MarkupCompiler.cs" />
+    <Compile Include="MarkupControlClass.cs" />
+    <Compile Include="MarkupControlInstance.cs" />
+    <Compile Include="MarkupElementCollection.cs" />
+    <Compile Include="MarkupFileCompiler.cs" />
+    <Compile Include="MarkupHtmlElement.cs" />
+    <Compile Include="MarkupNode.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\qb\qb.csproj">
+      <Project>{AD5371DD-34BC-42A5-B5EA-87FBFBFA5018}</Project>
+      <Name>qb</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\qc\qc.csproj">
+      <Project>{FDCDE62B-250B-4333-A8E6-55E49662D765}</Project>
+      <Name>qc</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Tests\comprehensive.qui" />
+    <EmbeddedResource Include="Tests\content.qui" />
+    <EmbeddedResource Include="Tests\entities.qui" />
+    <EmbeddedResource Include="Tests\simple.qui" />
+    <EmbeddedResource Include="Tests\simplehost.qui" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Tests\comprehensive.qui.css" />
+    <EmbeddedResource Include="Tests\comprehensive.qui.js" />
+    <EmbeddedResource Include="Tests\content.qui.js" />
+    <EmbeddedResource Include="Tests\entities.qui.js" />
+    <EmbeddedResource Include="Tests\simple.qui.js" />
+    <EmbeddedResource Include="Tests\simplehost.qui.js" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Tests\Sample\Bar.qui" />
+    <EmbeddedResource Include="Tests\Sample\Subfolder\Foo.qui" />
+  </ItemGroup>
+  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
+</Project>
