+++ tools/Tests/Tests/Sample/Sample.css
@@ -1,0 +1,12 @@
+/*** Foo ***/
+
+.Foo {
+    color: black;
+}
+
+/*** Bar ***/
+
+.Bar {
+    color: red;
+}
+
