+++ tools/Tests/Tests/simple.qui
@@ -1,0 +1,5 @@
+<!-- Simple control. -->
+
+<Control name="Simple">
+  <span id="Simple_content" />
+</Control>
