+++ tools/Tools.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 10.00
-# Visual C# Express 2008
+# Visual Studio 2008
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "qb", "qb\qb.csproj", "{AD5371DD-34BC-42A5-B5EA-87FBFBFA5018}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "qc", "qc\qc.csproj", "{FDCDE62B-250B-4333-A8E6-55E49662D765}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{0176E5C5-C12B-4386-8F23-4F4AD7462519}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,6 +21,10 @@ Global
 		{FDCDE62B-250B-4333-A8E6-55E49662D765}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FDCDE62B-250B-4333-A8E6-55E49662D765}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FDCDE62B-250B-4333-A8E6-55E49662D765}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0176E5C5-C12B-4386-8F23-4F4AD7462519}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0176E5C5-C12B-4386-8F23-4F4AD7462519}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0176E5C5-C12B-4386-8F23-4F4AD7462519}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0176E5C5-C12B-4386-8F23-4F4AD7462519}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
