+++ tools/qb/Application.cs
@@ -2,10 +2,6 @@
 using System.Collections.Generic;
 using System.IO;
 
-#if DEBUG
-using NUnit.Framework;
-#endif
-
 namespace qb
 {
     class Application
