--- tools/qb/Tests/Sample/Bar.qui
@@ -1,15 +1,0 @@
-<!-- Subclass -->
-
-<Control name="Bar">
-
-<prototype>
-    <Foo>Hello, world.</Foo>
-</prototype>
-
-<style>
-{
-    color: red;
-}
-</style>
-
-</Control>
