--- tools/qb/Tests/Sample/Sample.css
@@ -1,12 +1,0 @@
-/*** Foo ***/
-
-.Foo {
-    color: black;
-}
-
-/*** Bar ***/
-
-.Bar {
-    color: red;
-}
-
