+++ tools/qc/HtmlElementNames.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Xml.Linq;
-
-#if DEBUG
-using NUnit.Framework;
-#endif
 
 namespace qc
 {
@@ -16,7 +11,7 @@ namespace qc
             {
                 if (elements == null)
                 {
-                    elements = new ArrayList(Utilities.DeserializeXml<string[]>("qc.HtmlElementNames.xml"));
+                    elements = new ArrayList(Utilities.DeserializeXml<string[]>("qc.HtmlElementNames.xml", typeof(HtmlElementNames).Assembly));
                 }
 
                 return elements;
@@ -35,19 +30,5 @@ namespace qc
                 && element.Name != null
                 && IsHtmlElement(element.Name.ToString());
         }
-
-#if DEBUG
-        [TestFixture]
-        public class Tests
-        {
-            [Test]
-            public void CheckElements()
-            {
-                Assert.IsTrue(IsHtmlElement("div"));
-                Assert.IsTrue(IsHtmlElement("table"));
-                Assert.IsFalse(IsHtmlElement("foo"));
-            }
-        }
-#endif
     }
 }
