+++ tools/qc/MarkupElementCollection.cs
@@ -5,10 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Xml.Linq;
 
-#if DEBUG
-using NUnit.Framework;
-#endif
-
 namespace qc
 {
     /// <summary>
@@ -72,39 +68,5 @@ namespace qc
         {
             return Items.GetEnumerator();
         }
-
-#if DEBUG
-        [TestFixture]
-        public new class Tests
-        {
-            [Test]
-            public void Collection()
-            {
-                XElement element = new XElement("Foo",
-                    new XElement("Bar",
-                        new XAttribute("id", "bar"),
-                        "Control content"),
-                    new XElement("p", "paragraph")
-                );
-                MarkupNode node = MarkupNode.Parse(element.Nodes());
-                Assert.IsInstanceOf<MarkupElementCollection>(node);
-                MarkupElementCollection collection = (MarkupElementCollection) node;
-                Assert.IsNotNull(collection);
-                Assert.AreEqual(2, collection.Count());
-                List<MarkupElement> items = new List<MarkupElement>(collection);
-                Assert.IsInstanceOf<MarkupControlInstance>(items[0]);
-                Assert.IsInstanceOf<MarkupHtmlElement>(items[1]);
-                Assert.AreEqual(
-                    "[\n" +
-                    "\tthis.bar = QuickUI.Control.create(Bar, {\n" +
-                    "\t\t\"content\": \"Control content\",\n" +
-                    "\t\t\"id\": \"bar\"\n" +
-                    "\t}),\n" +
-                    "\t\"<p>paragraph</p>\"\n" +
-                    "]",
-                    node.JavaScript());
-            }
-        }
-#endif
     }
 }
