--- tools/qc/Tests/comprehensive.qui
@@ -1,43 +1,0 @@
-<!-- Comprehensive control tests a bunch of compiler features. -->
-
-<Control name="Comprehensive" tag="div">
-
-<prototype>
-  <SampleBaseClass>
-    <Simple id="message">
-      Hello, <i>world</i>!
-    </Simple>
-    <div>
-      <p id="Comprehensive_content" />
-    </div>
-  </SampleBaseClass>
-</prototype>
-
-<style>
-
-/* Base control style */
-{
-    font-family: "Arial";
-}
-
-p {
-    font-family: "Times New Roman";
-}
-
-.Comprehensive.combo {
-    color: red;
-}
-
-</style>
-
-<script>
-$.extend(Comprehensive.prototype, {
-  content: QuickUI.Element("Comprehensive_content").content(),
-  foo: function()
-  {
-      alert("Hello, world!");
-  }
-});
-</script>
-
-</Control>
