--- tools/qc/Tests/entities.qui.js
@@ -1,15 +1,0 @@
-//
-// Entities
-//
-Entities = QuickUI.Control.extend({
-	className: "Entities"
-});
-$.extend(Entities.prototype, {
-  ready: function() {
-    if (1 < 2 && 4 > 3)
-    {
-      alert("Success");
-    }
-  }
-});
-
