+++ tools/qc/qc.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>9.0.30729</ProductVersion>
+    <ProductVersion>9.0.21022</ProductVersion>
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectGuid>{FDCDE62B-250B-4333-A8E6-55E49662D765}</ProjectGuid>
     <OutputType>Exe</OutputType>
@@ -34,9 +34,6 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="nunit.framework, Version=2.5.3.9345, Culture=neutral, PublicKeyToken=96d09a1eb7f44a77, processorArchitecture=MSIL">
-      <SpecificVersion>False</SpecificVersion>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core">
       <RequiredTargetFramework>3.5</RequiredTargetFramework>
@@ -66,28 +63,7 @@
     <Compile Include="Utilities.cs" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Tests\simple.qui" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Tests\comprehensive.qui" />
-    <EmbeddedResource Include="Tests\content.qui" />
-    <EmbeddedResource Include="Tests\simplehost.qui" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Tests\comprehensive.qui.js" />
-    <EmbeddedResource Include="Tests\content.qui.js" />
-    <EmbeddedResource Include="Tests\simple.qui.js" />
-    <EmbeddedResource Include="Tests\simplehost.qui.js" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Tests\comprehensive.qui.css" />
-  </ItemGroup>
-  <ItemGroup>
     <None Include="app.config" />
-    <EmbeddedResource Include="Tests\entities.qui" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Tests\entities.qui.js" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="HtmlElementNames.xml" />
